Reorganize xbox controller script locations
